dreamobjects: make demo.py usable on sharedserver.
Dreamhost VPS & shared server don't have the ability to install
packages, but they DO have boto already installed.

The demo tries to make use of termcolor to look pretty, but that's not
available on older platforms, and isn't actually needed for the demo to
work; so fall back to not using colors if termcolor is not present.

Demo now confirmed to work as non-root on shared server & VPS.

Fixes: https://jira.dreamhost.com/browse/DHO-1017
Signed-off-by: Robin H. Johnson <robin.johnson@dreamhost.com>