Using Figure captions instead of lists in documentation
Lists, ordered and unordered, seem to create issues with zendesk
rendering of the output html. Instead of lists, it's preferrable
to use image caption in the .. figure:: directive.

This document is a first attempt at using this tecnique. The output
is also compliant with styleguide where they require the image to be
shown *before* the text describing the actions to take.