Performed further edits to Cloud docs style guide
These are the remaining edits to the style guide that Caleb and I have
been working on. Mostly punctuation fixes, changing the tone for a
public (non-internal) audience, and changed a few words. I will be
adding some more edits later to this document for how to handle images,
such as size, border, position, etc. (per Justin Lund). Let me know how
this commit looks, as I'm sorta a newb at this.